nexus-lm/requirements.txt
torch>=2.1.0
sentencepiece>=0.1.99
datasets>=2.14.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
pytest>=7.4.0
lm-eval>=0.4.0
wandb>=0.16.0
